+++ inst/doc/z-advanced-topic-walkthrough-project.R
@@ -1,0 +1,24 @@
+## ----include = FALSE----------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+## ----include=FALSE------------------------------------------------------------
+library(rix)
+
+## ----eval=FALSE---------------------------------------------------------------
+# library(rix)
+# 
+# # Choose the path to your project
+# path_default_nix <- "."
+# 
+# rix(
+#   date = "2025-02-28", # We recommend using a date
+#   r_pkgs = c("dplyr", "ggplot2"), # List all the packages you need
+#   ide = "rstudio", # List whatever editor you need
+#   project_path = path_default_nix,
+#   overwrite = TRUE,
+#   print = FALSE
+# )
+