+++ tests/testthat/test-fetchers.R
@@ -1,5 +1,6 @@
 testthat::test_that("Test fetchgit works", {
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
   testthat::expect_equal(
     fetchgit(
       list(
@@ -21,12 +22,14 @@ testthat::test_that("Test fetchgit fails gracefully", {
         repo_url = "https://github.com/rap4all/housing666/",
         commit = "this_commit_is_wrong"
       )
-    ), "Are these correct?"
+    ),
+    "Are these correct?"
   )
 })
 
 testthat::test_that("Test fetchgit works with gitlab packages", {
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
   testthat::expect_equal(
     fetchgit(
       list(
@@ -41,6 +44,7 @@ testthat::test_that("Test fetchgit works with gitlab packages", {
 
 testthat::test_that("Test fetchgit works with packages with empty imports", {
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
   testthat::expect_equal(
     fetchgit(
       list(
@@ -55,6 +59,7 @@ testthat::test_that("Test fetchgit works with packages with empty imports", {
 
 testthat::test_that("Test fetchzip works", {
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
   testthat::expect_equal(
     fetchzip("AER@1.2-8"),
     "\n    AER = (pkgs.rPackages.buildRPackage {\n      name = \"AER\";\n      src = pkgs.fetchzip {\n       url = \"https://cran.r-project.org/src/contrib/Archive/AER/AER_1.2-8.tar.gz\";\n       sha256 = \"sha256-OqxXcnUX/2C6wfD5fuNayc8OU+mstI3tt4eBVGQZ2S0=\";\n      };\n      propagatedBuildInputs = builtins.attrValues {\n        inherit (pkgs.rPackages) \n          car\n          lmtest\n          sandwich\n          survival\n          zoo\n          Formula;\n      };\n    });\n"
@@ -71,23 +76,26 @@ testthat::test_that("Test fetchzip fails gracefully", {
 
 testthat::test_that("Test fetchgits", {
   testthat::skip_on_cran()
-  testthat::expect_equal(
-    fetchgits(
-      list(
-        list(
-          package_name = "housing",
-          repo_url = "https://github.com/rap4all/housing/",
-          commit = "1c860959310b80e67c41f7bbdc3e84cef00df18e"
-        ),
-        list(
-          package_name = "fusen",
-          repo_url = "https://github.com/ThinkR-open/fusen",
-          commit = "d617172447d2947efb20ad6a4463742b8a5d79dc"
-        )
-      )
+  skip_if_not(nix_shell_available())
+  cache_file <- get_cache_file()
+  pkg_list <- list(
+    list(
+      package_name = "housing",
+      repo_url = "https://github.com/rap4all/housing/",
+      commit = "1c860959310b80e67c41f7bbdc3e84cef00df18e"
     ),
-    "\n    fusen = (pkgs.rPackages.buildRPackage {\n      name = \"fusen\";\n      src = pkgs.fetchgit {\n        url = \"https://github.com/ThinkR-open/fusen\";\n        rev = \"d617172447d2947efb20ad6a4463742b8a5d79dc\";\n        sha256 = \"sha256-TOHA1ymLUSgZMYIA1a2yvuv0799svaDOl3zOhNRxcmw=\";\n      };\n      propagatedBuildInputs = builtins.attrValues {\n        inherit (pkgs.rPackages) \n          attachment\n          cli\n          desc\n          devtools\n          glue\n          here\n          magrittr\n          parsermd\n          roxygen2\n          stringi\n          tibble\n          tidyr\n          usethis\n          yaml;\n      };\n    });\n\n\n    housing = (pkgs.rPackages.buildRPackage {\n      name = \"housing\";\n      src = pkgs.fetchgit {\n        url = \"https://github.com/rap4all/housing/\";\n        rev = \"1c860959310b80e67c41f7bbdc3e84cef00df18e\";\n        sha256 = \"sha256-s4KGtfKQ7hL0sfDhGb4BpBpspfefBN6hf+XlslqyEn4=\";\n      };\n      propagatedBuildInputs = builtins.attrValues {\n        inherit (pkgs.rPackages) \n          dplyr\n          ggplot2\n          janitor\n          purrr\n          readxl\n          rlang\n          rvest\n          stringr\n          tidyr;\n      };\n    });\n"
+    list(
+      package_name = "fusen",
+      repo_url = "https://github.com/ThinkR-open/fusen",
+      commit = "d617172447d2947efb20ad6a4463742b8a5d79dc"
+    )
   )
+  expected_output <- "\n    fusen = (pkgs.rPackages.buildRPackage {\n      name = \"fusen\";\n      src = pkgs.fetchgit {\n        url = \"https://github.com/ThinkR-open/fusen\";\n        rev = \"d617172447d2947efb20ad6a4463742b8a5d79dc\";\n        sha256 = \"sha256-TOHA1ymLUSgZMYIA1a2yvuv0799svaDOl3zOhNRxcmw=\";\n      };\n      propagatedBuildInputs = builtins.attrValues {\n        inherit (pkgs.rPackages) \n          attachment\n          cli\n          desc\n          devtools\n          glue\n          here\n          magrittr\n          parsermd\n          roxygen2\n          stringi\n          tibble\n          tidyr\n          usethis\n          yaml;\n      };\n    });\n\n\n    housing = (pkgs.rPackages.buildRPackage {\n      name = \"housing\";\n      src = pkgs.fetchgit {\n        url = \"https://github.com/rap4all/housing/\";\n        rev = \"1c860959310b80e67c41f7bbdc3e84cef00df18e\";\n        sha256 = \"sha256-s4KGtfKQ7hL0sfDhGb4BpBpspfefBN6hf+XlslqyEn4=\";\n      };\n      propagatedBuildInputs = builtins.attrValues {\n        inherit (pkgs.rPackages) \n          dplyr\n          ggplot2\n          janitor\n          purrr\n          readxl\n          rlang\n          rvest\n          stringr\n          tidyr;\n      };\n    });\n"
+  testthat::expect_equal(
+    fetchgits(pkg_list),
+    expected_output
+  )
+  on.exit(unlink(cache_file))
 })
 
 testthat::test_that("Test fetchgits works when PR is provided in a remote package, but does not use it", {
@@ -95,6 +103,8 @@ testthat::test_that("Test fetchgits works when PR is provided in a remote packag
   # see https://github.com/mihem/rixTest/commit/b56829f7771d131e02fc58c546f9af6ee13b857e
   # This should not fail, however it will not use the PR and instead fetch the closest commit using GitHub API
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
+  cache_file <- get_cache_file()
   pkg_list <- list(
     package_name = "rixTest",
     repo_url = "https://github.com/mihem/rixTest",
@@ -132,6 +142,7 @@ testthat::test_that("Test fetchgits works when PR is provided in a remote packag
     fetchgits(pkg_list),
     expected_output
   )
+  on.exit(unlink(cache_file))
 })
 
 testthat::test_that("Test fetchgits works when tag is provided in a remote package, but does not use it", {
@@ -139,6 +150,8 @@ testthat::test_that("Test fetchgits works when tag is provided in a remote packa
   # https://github.com/mihem/rixTest/commit/25da90697895b006934a70bbd003aab5c5206c8b
   # This should not fail, however it will not use the tag and instead fetch the closest commit using GitHub API
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
+  cache_file <- get_cache_file()
   pkg_list <- list(
     list(
       package_name = "rixTest",
@@ -153,10 +166,12 @@ testthat::test_that("Test fetchgits works when tag is provided in a remote packa
     fetchgits(pkg_list),
     expected_output
   )
+  on.exit(unlink(cache_file))
 })
 
 testthat::test_that("Test fetchzips works", {
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
   testthat::expect_equal(
     fetchzips(
       c("dplyr@0.8.0", "AER@1.2-8")
@@ -167,6 +182,7 @@ testthat::test_that("Test fetchzips works", {
 
 testthat::test_that("Test fetchpkgs works", {
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
   testthat::expect_equal(
     fetchpkgs(
       git_pkgs = list(
@@ -189,6 +205,7 @@ testthat::test_that("Test fetchpkgs works", {
 
 testthat::test_that("Test fetchgit gets a package with several remote deps and commits", {
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
   testthat::expect_equal(
     suppressMessages(
       fetchgit(
@@ -205,6 +222,7 @@ testthat::test_that("Test fetchgit gets a package with several remote deps and c
 
 testthat::test_that("Test fetchgit gets a package that is not listed in DESCRIPTION, only in NAMESPACE", {
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
   testthat::expect_equal(
     fetchgit(
       list(
@@ -219,6 +237,7 @@ testthat::test_that("Test fetchgit gets a package that is not listed in DESCRIPT
 
 testthat::test_that("Test fetchgit works even if there are not `importfrom` in NAMESPACE", {
   testthat::skip_on_cran()
+  skip_if_not(nix_shell_available())
   testthat::expect_no_error(
     fetchgit(
       list(
@@ -312,7 +331,11 @@ testthat::test_that("resolve_package_commit works with different input cases", {
   remotes <- c("welch-lab/liger", "SaskiaFreytag/schex@031320d")
   target_date <- "2024-04-04T14:16:11Z"
   testthat::expect_equal(
-    resolve_package_commit(remote_pkg_name_and_ref = pkg_with_ref, date = target_date, remotes = remotes),
+    resolve_package_commit(
+      remote_pkg_name_and_ref = pkg_with_ref,
+      date = target_date,
+      remotes = remotes
+    ),
     "031320d"
   )
 
@@ -321,7 +344,11 @@ testthat::test_that("resolve_package_commit works with different input cases", {
   remotes <- c("welch-lab/liger", "hms-dbmi/conos")
   target_date <- "2024-04-04T14:16:11Z"
   testthat::expect_equal(
-    resolve_package_commit(remote_pkg_name_and_ref = pkg_without_ref, date = target_date, remotes = remotes),
+    resolve_package_commit(
+      remote_pkg_name_and_ref = pkg_without_ref,
+      date = target_date,
+      remotes = remotes
+    ),
     "43fccb96b986f9da2c3a4320fe58693ca660193b"
   )